core: Move api::NotBefore into types
All other types that are used in the API are in types, so this patch
also moves api::NotBefore there.
